Link Objects to Collections operator

+++ src/r0tools_simple_toolbox/utils/general.py
@@ -289,18 +289,20 @@ def collections_create_new(name: str):
     return collection
 
 
-def collection_link_object(collection, obj):
+def collection_link_object(collection, obj, unlink_others: bool = False):
     # Sometimes, the reference can be NoneType.
     if not obj:
         return False
 
-    for coll in bpy.data.collections:
-        if obj.name in coll.objects:
-            coll.objects.unlink(obj)
+    if unlink_others:
+        for coll in bpy.data.collections:
+            if obj.name in coll.objects:
+                coll.objects.unlink(obj)
 
     # Loop for user collections that object is linked to
-    for coll in obj.users_collection:
-        coll.objects.unlink(obj)
+    if unlink_others:
+        for coll in obj.users_collection:
+            coll.objects.unlink(obj)
 
     collection.objects.link(obj)
